+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+v0.8.4:
+  * Fix compatibility with Rust stable.
+
 v0.8.3
   * Allow manual rendering back-end selection with the `WGPU_BACKEND` environment variables.
   * Allow manual GPU selection with the `WGPU_POWER_PREF` environment variables.